Run CREATE2 unit tests for both LegacyVM and aleth-interpreter
